move feature file into feature folder
